+++ src/Atata.HtmlValidation/HtmlValidationOptions.cs
@@ -101,10 +101,23 @@ namespace Atata.HtmlValidation
         public Encoding Encoding { get; set; }
 
         /// <summary>
-        /// Gets or sets the recommended version of "html-validate" NPM package to install when there is no an installed one.
-        /// The default value is <c>"5.x"</c>.
+        /// Gets or sets the recommended version of "html-validate" NPM package to install.
+        /// The default value is <c>"5.1.1"</c>.
         /// </summary>
-        public string RecommendedHtmlValidatePackageVersion { get; set; } = "5.x";
+        [Obsolete("Use " + nameof(HtmlValidatePackageVersion) + " instead.")] // Obsolete since v1.2.0.
+        public string RecommendedHtmlValidatePackageVersion
+        {
+            get => HtmlValidatePackageVersion;
+            set => HtmlValidatePackageVersion = value;
+        }
+
+        /// <summary>
+        /// Gets or sets the required version of "html-validate" NPM package.
+        /// The required version will be installed if "html-validate" package is not installed or the installed version differs from the required one.
+        /// The default value is <c>"5.1.1"</c>.
+        /// Set <see langword="null"/> to disable the version check and use any pre-installed version.
+        /// </summary>
+        public string HtmlValidatePackageVersion { get; set; } = "5.1.1";
 
         /// <summary>
         /// Clones this instance with executing the action that can change some properties of the copy.